stage_II/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=0.4.6=pypi_0
apex=0.1=pypi_0
blas=2.111=mkl
blas-devel=3.9.0=11_win64_mkl
certifi=2021.5.30=pypi_0
charset-normalizer=2.0.4=pypi_0
colorama=0.4.4=pypi_0
cudatoolkit=10.2.89=hb195166_8
cycler=0.10.0=pypi_0
dataclasses=0.8=pyh787bdff_2
decorator=4.4.2=pypi_0
freetype=2.10.4=h546665d_1
geffnet=0.9.8=pypi_0
idna=3.2=pypi_0
imageio=2.9.0=pypi_0
imgaug=0.4.0=pypi_0
intel-openmp=2021.3.0=h57928b3_3372
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
kaggle=1.5.6=pypi_0
kiwisolver=1.3.1=pypi_0
libblas=3.9.0=11_win64_mkl
libcblas=3.9.0=11_win64_mkl
liblapack=3.9.0=11_win64_mkl
liblapacke=3.9.0=11_win64_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.2.0=hd0e1b90_0
libuv=1.42.0=h8ffe710_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.3.4=pypi_0
mkl=2021.3.0=hb70f87d_564
mkl-devel=2021.3.0=h57928b3_565
mkl-include=2021.3.0=hb70f87d_564
msys2-conda-epoch=20160418=1
munch=2.5.0=pypi_0
networkx=2.5.1=pypi_0
ninja=1.10.2=h5362a0b_0
nose=1.3.7=pypi_0
numpy=1.18.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.4.0.42=pypi_0
opencv-python-headless=4.5.3.56=pypi_0
pandas=0.24.2=pypi_0
pillow=7.2.0=pypi_0
pip=21.2.4=pyhd8ed1ab_0
pretrainedmodels=0.7.4=pypi_0
pyparsing=2.4.7=pypi_0
python=3.6.13=h39d44d4_0_cpython
python-dateutil=2.8.2=pypi_0
python-slugify=5.0.2=pypi_0
python_abi=3.6=2_cp36m
pytorch=1.9.0=py3.6_cuda10.2_cudnn7_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
requests=2.26.0=pypi_0
resnest=0.0.6b20200823=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=57.4.0=py36ha15d459_0
shapely=1.7.1=pypi_0
six=1.16.0=pypi_0
tbb=2021.3.0=h2d74725_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.2.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h8ffe710_0
torchaudio=0.9.0=py36
torchvision=0.10.0=py36_cu102
tqdm=4.31.1=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=3.10.0.0=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.24.3=pypi_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wheel=0.37.0=pyhd8ed1ab_1
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
